Fix bug in HTML rendering in Node.js
When supplied with a Cheerio instance (instead of a string of HTML), the
Node.js-specific `html` and `append` methods should reference the
complete outer HTML (not just the contents) as generated by Cheerio's
`html` method.

Note that this was previously fixed with d60cf12, but Cheerio's public
API changed, which forced reversion to the earlier (incorrect)
implementation at 3fb562160e.